backend/requirements.txt
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
fastapi==0.105.0
requests==2.31.0
selenium==4.17.2
coverage==7.4.1
flake8==7.0.0
mccabe==0.7.0
pycodestyle==2.11.1 
pyflakes==3.2.0
uvicorn==0.27.0.post1
alembic==1.13.1
SQLAlchemy==2.0.25
typing==3.7.4.3